+++ deep-capitalizer/src/components/Home.js
@@ -5,13 +5,13 @@ import { Link } from 'react-router-dom'
 const Home = () => {
   return (
     <>
-        <Link to='/testPage'>
-          <Button>Goto Visualization page</Button>
-        </Link>
-        <Link to='/app'>
-          <Button>Goto Server Test page</Button>
-        </Link>    
-        <p>This is Home</p>
+      <Link to='/visualization'>
+        <Button>Goto Visualization page</Button>
+      </Link>
+      <Link to='/server-test'>
+        <Button>Goto Server Test page</Button>
+      </Link>    
+      <p>This is Home</p>
     </>
   );
 }
