+++ deep-capitalizer/src/components/Visualization.js
@@ -1,13 +1,11 @@
 import './App.css';
-import logo from '../logo.svg';
 
 const Visualization = () => {
     return (
         <div className='App'>
           <header className='App-header'>
-            <img src={logo} className='App-logo' alt='logo' />
             <p>
-              Graphs go here
+              - Graphs go here -
             </p>
           </header>
         </div>
